refactor: 요청 데이터 생성 로직을 두 개의 클래스로 분리
